added styles
