+++ product.go
@@ -6,20 +6,6 @@ import (
 	"strings"
 )
 
-type Categories struct {
-	Accessories  Products `json:"Accessories"`
-	Bags         Products `json:"Bags"`
-	Skate        Products `json:"Skate"`
-	TopsSweaters Products `json:"Tops/Sweaters"`
-	Pants        Products `json:"Pants"`
-	Jackets      Products `json:"Jackets"`
-	Hats         Products `json:"Hats"`
-	Sweatshirts  Products `json:"Sweatshirts"`
-	Shirts       Products `json:"Shirts"`
-	Shorts       Products `json:"Shorts"`
-	New          Products `json:"new"`
-}
-
 type Product struct {
 	Name         string `json:"name"`
 	ID           int    `json:"id"`
